testing why profile is not rendering backup

+++ controllers/posts_controller.js
@@ -9,6 +9,17 @@ module.exports.create = async function (req, res) {
             content: req.body.content,
             user: req.user._id
         });
+
+        if(req.xhr){
+            return res.status(200).json({
+                data:{
+                    post:post
+                },
+                message:"Post created!"
+            })
+        }
+
+
         req.flash('success','Post published')
         //? Post creation successful
         return res.redirect('back');
@@ -42,6 +53,15 @@ module.exports.destroy = async function (req, res) {
         if (post.user == req.user.id) {
             // Use `await` with `post.remove` to delete the post
             await post.deleteOne();
+
+            if(req.xhr){
+                return res.status(200).json({
+                    data:{
+                        post_id:req.params.id
+                    },
+                    message:"Post deleted"
+                });
+            }
 
             // Use `await` with `Comment.deleteMany` to remove associated comments
             await Comment.deleteMany({ post: req.params.id });
